Fix 'broken' symbols when displaying emoji before them.
There's an issue with either mpv or libass where symbols are displayed
in squares when an emoji appears before them.

This doesn't occur if there is at least one symbol before the emoji. The
issue also seems to be font-dependent (the default OSD font is
affected):

show-text "🥝 ↪"
show-text "↪ 🥝 ↪"

This workaround adds an invisible character before the list to prevent
this issue.